Initial tool addition
The goal of this tool is to be able to read & write fw files to the
FT5x06 touch controllers.

This was largely inspired from the Focaltech Github driver:
https://github.com/focaltech-systems/drivers-input-touchscreen-FTS_driver

From there hit & trial produced this piece of code.

So far only writing the FW has been tested on a Boundary Tianma display
which includes a FT5426 touch controller:
https://boundarydevices.com/product/bd070lic2/

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>